Added proto definition file for grpc api Refactored all existing code to use transaction struct generated by protobuf All tests passing

+++ pkg/query/query_test.go
@@ -16,7 +16,7 @@ func TestWalletBalance(t *testing.T) {
 	t.Run("should return the balance for a given wallet ID", func(t *testing.T) {
 		book := memory.NewMockInMemoryBook()
 		wallet := "1"
-		expected := 100000
+		expected := int32(100000)
 
 		balance, err := WalletBalance(book, wallet)
